fix illegal memory access in attention

+++ aten/src/ATen/native/transformers/cuda/mem_eff_attention/kernel_backward.h
@@ -636,8 +636,8 @@ struct AttentionBackwardKernel {
     const scalar_t* output_ptr = nullptr; // [Mq, nH, Kv]
     const scalar_t* grad_output_ptr = nullptr; // [Mq, nH, Kv]
     accum_t* delta_ptr = nullptr; // [nH, Mq]
-    const int32_t* cu_seqlens_q_ptr = nullptr;
-    const int32_t* cu_seqlens_k_ptr = nullptr;
+    const int64_t* cu_seqlens_q_ptr = nullptr;
+    const int64_t* cu_seqlens_k_ptr = nullptr;
 
     // Output tensors
     output_t* grad_query_ptr = nullptr; //  [Mq, nH, K]
@@ -659,19 +659,19 @@ struct AttentionBackwardKernel {
     accum_t scale = 1.0f;
 
     // Dimensions/strides
-    int32_t head_dim = -1;
-    int32_t head_dim_value = -1;
-    int32_t num_queries = -1;
-    int32_t num_keys = -1;
-    int32_t num_heads = -1;
+    int64_t head_dim = -1;
+    int64_t head_dim_value = -1;
+    int64_t num_queries = -1;
+    int64_t num_keys = -1;
+    int64_t num_heads = -1;
     uint8_t custom_mask_type = NoCustomMask;
 
-    int32_t q_strideM = -1;
-    int32_t k_strideM = -1;
-    int32_t v_strideM = -1;
-    int32_t bias_strideM = 0;
-    int32_t gO_strideM = -1;
-    int32_t gB_strideM = -1;
+    int64_t q_strideM = -1;
+    int64_t k_strideM = -1;
+    int64_t v_strideM = -1;
+    int64_t bias_strideM = 0;
+    int64_t gO_strideM = -1;
+    int64_t gB_strideM = -1;
     int8_t gQKV_strideM_multiplier = 1; // 3 for packed, 1 otherwise
 
     at::PhiloxCudaState rng_engine_inputs = {0, 0};
@@ -680,16 +680,16 @@ struct AttentionBackwardKernel {
     unsigned long long dropout_batch_head_rng_offset = 0;
     float dropout_prob = 0.0f;
 
-    CUTLASS_HOST_DEVICE int32_t o_strideM() const {
+    CUTLASS_HOST_DEVICE int64_t o_strideM() const {
       return head_dim_value * num_heads;
     }
-    CUTLASS_HOST_DEVICE int32_t gQ_strideM() const {
+    CUTLASS_HOST_DEVICE int64_t gQ_strideM() const {
       return gQKV_strideM_multiplier * num_heads * head_dim;
     }
-    CUTLASS_HOST_DEVICE int32_t gK_strideM() const {
+    CUTLASS_HOST_DEVICE int64_t gK_strideM() const {
       return gQKV_strideM_multiplier * num_heads * head_dim;
     }
-    CUTLASS_HOST_DEVICE int32_t gV_strideM() const {
+    CUTLASS_HOST_DEVICE int64_t gV_strideM() const {
       return gQKV_strideM_multiplier * num_heads * head_dim_value;
     }
 
@@ -772,8 +772,8 @@ struct AttentionBackwardKernel {
         assert(cu_seqlens_k_ptr != nullptr);
         cu_seqlens_q_ptr += batch_id;
         cu_seqlens_k_ptr += batch_id;
-        int32_t q_start = cu_seqlens_q_ptr[0];
-        int32_t k_start = cu_seqlens_k_ptr[0];
+        int64_t q_start = cu_seqlens_q_ptr[0];
+        int64_t k_start = cu_seqlens_k_ptr[0];
         int64_t q_next_start = cu_seqlens_q_ptr[1];
         int64_t k_next_start = cu_seqlens_k_ptr[1];
         assert(q_next_start - q_start <= num_queries);
@@ -858,21 +858,21 @@ struct AttentionBackwardKernel {
         return 0;
       }
       return num_splits_key * kBlockSizeJ *
-          align_up(head_dim, (int32_t)kBlockSizeI);
+          align_up(head_dim, (int64_t)kBlockSizeI);
     }
     CUTLASS_HOST_DEVICE int64_t workspace_elements_gv() const {
       if (!kNeedsAccumGradV) {
         return 0;
       }
       return num_splits_key * kBlockSizeJ *
-          align_up(head_dim_value, (int32_t)kBlockSizeI);
+          align_up(head_dim_value, (int64_t)kBlockSizeI);
     }
     CUTLASS_HOST_DEVICE int64_t workspace_elements_gq() const {
       if (!kNeedsAccumGradQ) {
         return 0;
       }
-      int num_blocks = ceil_div(num_queries, kBlockSizeI);
-      int num_cols = ceil_div(head_dim, MatmulGradQ::ThreadblockShape::kN);
+      int num_blocks = ceil_div(num_queries, (int64_t)kBlockSizeI);
+      int num_cols = ceil_div(head_dim, (int64_t)MatmulGradQ::ThreadblockShape::kN);
       return num_blocks * num_cols * sizeof(GradQTempStorage) /
           sizeof(output_accum_t);
     }
@@ -1286,7 +1286,7 @@ struct AttentionBackwardKernel {
     TORCH_CHECK(
         p.num_splits_key > 0, "Invalid `num_splits_key` (expected >0)");
     TORCH_CHECK(
-        p.num_splits_key <= cutlass::ceil_div(p.num_keys, kBlockSizeJ),
+        p.num_splits_key <= cutlass::ceil_div(p.num_keys, (int64_t)kBlockSizeJ),
         "Invalid `num_splits_key` (",
         p.num_splits_key,
         ") - too large for `num_keys` = ",
@@ -1307,12 +1307,12 @@ struct AttentionBackwardKernel {
     uint8_t warp_id = warp_uniform(thread_id / 32);
     uint8_t lane_id = thread_id % 32;
 
-    int32_t key_start = p.split_key_device() * kBlockSizeJ;
+    int64_t key_start = p.split_key_device() * kBlockSizeJ;
     if (key_start >= p.num_keys) {
       return;
     }
     if (kPrologueQK) {
-      int32_t query_start = getQueryStart(p, key_start);
+      int64_t query_start = getQueryStart(p, key_start);
       prologueQkNextIteration<true>(
           shared_storage, p, query_start, key_start, warp_id, lane_id);
     }
@@ -1362,8 +1362,8 @@ struct AttentionBackwardKernel {
          key_start += p.num_splits_key_device() * kBlockSizeJ) {
       output_frags.clear();
 
-      int32_t next_key = key_start;
-      int32_t query_start = getQueryStart(p, key_start);
+      int64_t next_key = key_start;
+      int64_t query_start = getQueryStart(p, key_start);
       while (next_key == key_start && query_start < p.num_queries) {
         // This line here
         // vvvvvvvvvvvvvv
@@ -1385,7 +1385,7 @@ struct AttentionBackwardKernel {
             warp_id,
             lane_id);
 
-        int32_t next_query;
+        int64_t next_query;
         incrIteration(p, query_start, key_start, next_query, next_key);
         query_start = next_query;
       }
@@ -1439,8 +1439,8 @@ struct AttentionBackwardKernel {
       SharedStorage& shared_storage,
       OutputFragments& output_frags,
       Params& p,
-      int32_t query_start,
-      int32_t key_start,
+      int64_t query_start,
+      int64_t key_start,
       const curandStatePhilox4_32_10_t& curand_state_init,
       uint8_t warp_id,
       uint8_t lane_id) {
@@ -1463,7 +1463,7 @@ struct AttentionBackwardKernel {
     };
 
     bool isFirstQuery = (query_start == getQueryStart(p, key_start));
-    int32_t next_query, next_key;
+    int64_t next_query, next_key;
     incrIteration(p, query_start, key_start, next_query, next_key);
     bool isLastQuery = next_key != key_start;
 
@@ -1475,16 +1475,16 @@ struct AttentionBackwardKernel {
       shared_storage.di()[thread_id] = di_rf;
     }
 
-    int32_t num_queries_in_block = skipBoundsChecks
+    int64_t num_queries_in_block = skipBoundsChecks
         ? MatmulQK::Mma::Shape::kN
         : warp_uniform(cutlass::fast_min(
-              (int32_t)MatmulQK::Mma::Shape::kN, p.num_queries - query_start));
-    int32_t num_keys_in_block = skipBoundsChecks
+              (int64_t)MatmulQK::Mma::Shape::kN, p.num_queries - query_start));
+    int64_t num_keys_in_block = skipBoundsChecks
         ? MatmulQK::Mma::Shape::kM
         : warp_uniform(cutlass::fast_min(
-              (int32_t)MatmulQK::Mma::Shape::kM, p.num_keys - key_start));
+              (int64_t)MatmulQK::Mma::Shape::kM, p.num_keys - key_start));
 
-    auto prologueGradV = [&](int col) {
+    auto prologueGradV = [&](int64_t col) {
       typename MatmulGradV::Mma::IteratorB iterator_dO(
           {int32_t(p.gO_strideM)},
           const_cast<scalar_t*>(p.grad_output_ptr + query_start * p.gO_strideM + col),
@@ -1709,7 +1709,7 @@ struct AttentionBackwardKernel {
         // on the K-dimension) otherwise we can get NaNs during the GEMM
         const int kQueriesPerBlock = kBlockSizeI;
         const int threads_per_row = cutlass::fast_min(
-            int32_t(kNumThreads / kQueriesPerBlock), num_keys_in_block);
+            int64_t(kNumThreads / kQueriesPerBlock), num_keys_in_block);
         const int elts_per_thread = cutlass::round_nearest(
             cutlass::ceil_div(num_keys_in_block, threads_per_row), 4);
 
@@ -1779,7 +1779,7 @@ struct AttentionBackwardKernel {
     /////////////////////////////////////////////////////////////////////////////////////////////////
     constexpr bool kSingleIterationGradV =
         kMaxK <= MatmulGradV::ThreadblockShape::kN;
-    for (int col = 0; col < (kSingleIterationGradV ? 1 : p.head_dim_value);
+    for (int64_t col = 0; col < (kSingleIterationGradV ? 1 : p.head_dim_value);
          col += MatmulGradV::ThreadblockShape::kN) {
       using Mma = typename MatmulGradV::Mma;
       using AccumTileGmem = typename MatmulGradQ::AccumTileGmem;
@@ -2057,7 +2057,7 @@ struct AttentionBackwardKernel {
       int col_id = col / MatmulGradQ::ThreadblockShape::kN;
       int num_cols = kSingleIterationGradQ
           ? 1
-          : ceil_div(p.head_dim, MatmulGradQ::ThreadblockShape::kN);
+          : ceil_div(p.head_dim, static_cast<int64_t>(MatmulGradQ::ThreadblockShape::kN));
       int storage_id = (col_id + query_start / kBlockSizeI * num_cols);
 
       if (p.num_splits_key_device() > 1) {
@@ -2232,7 +2232,7 @@ struct AttentionBackwardKernel {
       }
 
       if (kPrologueQK && isLastColumn) {
-        int32_t next_query, next_key;
+        int64_t next_query, next_key;
         incrIteration(p, query_start, key_start, next_query, next_key);
         DISPATCH_BOOL(
             next_key != key_start, kForceReloadK, ([&]() {
@@ -2260,7 +2260,7 @@ struct AttentionBackwardKernel {
     }
   }
 
-  static CUTLASS_HOST_DEVICE int32_t getQueryStartShift(Params const& p) {
+  static CUTLASS_HOST_DEVICE int64_t getQueryStartShift(Params const& p) {
     if (p.custom_mask_type == NoCustomMask && p.num_splits_key_device() > 1) {
       return (p.split_key_device() * kBlockSizeI) % getQueryEnd(p);
     }
@@ -2268,27 +2268,27 @@ struct AttentionBackwardKernel {
   }
 
   // Iteration order logic
-  static CUTLASS_HOST_DEVICE int32_t
-  getQueryStart(Params const& p, int32_t key_start) {
+  static CUTLASS_HOST_DEVICE int64_t
+  getQueryStart(Params const& p, int64_t key_start) {
     return getSmallestQueryForKey(p, key_start) + getQueryStartShift(p);
   };
-  static CUTLASS_HOST_DEVICE int32_t getQueryEnd(Params const& p) {
-    return align_up(p.num_queries, kBlockSizeI);
+  static CUTLASS_HOST_DEVICE int64_t getQueryEnd(Params const& p) {
+    return align_up(p.num_queries, (int64_t)kBlockSizeI);
   };
 
-  static CUTLASS_HOST_DEVICE int32_t
-  getSmallestQueryForKey(Params const& p, int32_t key_start) {
+  static CUTLASS_HOST_DEVICE int64_t
+  getSmallestQueryForKey(Params const& p, int64_t key_start) {
     if (p.custom_mask_type == NoCustomMask) {
       return 0;
     }
-    int32_t shift = p.custom_mask_type == CausalFromBottomRight
+    int64_t shift = p.custom_mask_type == CausalFromBottomRight
         ? p.num_keys - p.num_queries
         : 0;
-    int32_t window_size =
+    int64_t window_size =
         p.window_size == 0 ? p.num_queries + p.num_keys : p.window_size;
 
     auto last_key_for_block =
-        cutlass::fast_min(key_start + kBlockSizeJ, p.num_keys) - 1;
+        cutlass::fast_min(key_start + (int64_t)kBlockSizeJ, p.num_keys) - 1;
     int first_query = key_start - shift;
     int last_query = last_key_for_block - shift + window_size - 1;
     if (last_query < 0 || first_query >= p.num_queries) {
@@ -2303,15 +2303,15 @@ struct AttentionBackwardKernel {
   // for instance in the causal case, or varying seqlen
   static CUTLASS_HOST_DEVICE int32_t
   getNumParallelBlocksForQuery(Params const& p, int32_t query_start) {
-    int16_t num_key_blocks = ceil_div(p.num_keys, kBlockSizeJ);
+    int16_t num_key_blocks = ceil_div(p.num_keys, static_cast<int64_t>(kBlockSizeJ));
     if (p.custom_mask_type != NoCustomMask) {
       int32_t shift = p.custom_mask_type == CausalFromBottomRight
           ? p.num_keys - p.num_queries
           : 0;
       int32_t last_query_for_block =
-          cutlass::fast_min(query_start + kBlockSizeI, p.num_queries) - 1;
+          cutlass::fast_min(static_cast<int64_t>(query_start + kBlockSizeI), p.num_queries) - 1;
       int32_t last_key_for_block =
-          cutlass::fast_min(last_query_for_block + shift, p.num_keys - 1);
+          cutlass::fast_min(static_cast<int64_t>(last_query_for_block + shift), p.num_keys - 1);
       int32_t first_key_for_block = p.window_size == 0
           ? 0
           : cutlass::fast_max(query_start - p.window_size + 1 + shift, 0);
@@ -2333,10 +2333,10 @@ struct AttentionBackwardKernel {
   // Returns the next block to process
   static CUTLASS_HOST_DEVICE void incrIteration(
       Params const& p,
-      int32_t query_start,
-      int32_t key_start,
-      int32_t& next_query,
-      int32_t& next_key) {
+      int64_t query_start,
+      int64_t key_start,
+      int64_t& next_query,
+      int64_t& next_key) {
     next_query = query_start + kBlockSizeI;
     next_key = key_start;
     auto query_shift = getQueryStartShift(p);
@@ -2357,7 +2357,7 @@ struct AttentionBackwardKernel {
             : 0;
         // last key that is not masked out
         int last_key_for_block =
-            cutlass::fast_min(key_start + kBlockSizeJ, p.num_keys) - 1;
+            cutlass::fast_min(key_start + (int64_t)kBlockSizeJ, p.num_keys) - 1;
         int last_query = last_key_for_block - shift + p.window_size - 1;
         if (next_query <= last_query && next_query < p.num_queries) {
           return;
@@ -2368,7 +2368,7 @@ struct AttentionBackwardKernel {
       // jump to next key
     }
     // Next key
-    next_key = key_start + p.num_splits_key_device() * kBlockSizeJ;
+    next_key = key_start + p.num_splits_key_device() * (int64_t)kBlockSizeJ;
     next_query = getQueryStart(p, next_key);
   }
 
@@ -2422,11 +2422,11 @@ struct AttentionBackwardKernel {
     int32_t num_keys_in_block = skipBoundsChecks
         ? MatmulQK::Mma::Shape::kM
         : cutlass::fast_min(
-              (int32_t)MatmulQK::Mma::Shape::kM, p.num_keys - key_start);
+              (int64_t)MatmulQK::Mma::Shape::kM, p.num_keys - key_start);
     typename MatmulGradV::OutputTileIterator outputV_it(
         typename MatmulGradV::OutputTileIterator::Params{p.gV_strideM()},
         p.grad_value_ptr + key_start * p.gV_strideM(),
-        {num_keys_in_block, p.head_dim_value},
+        {(int64_t)num_keys_in_block, p.head_dim_value},
         thread_id);
     accumulateInGmem<MatmulGradV>(
         shared_storage.gradV_epilogue_final(),
@@ -2439,7 +2439,7 @@ struct AttentionBackwardKernel {
     typename MatmulGradK::OutputTileIterator outputK_it(
         typename MatmulGradK::OutputTileIterator::Params{p.gK_strideM()},
         p.grad_key_ptr + key_start * p.gK_strideM(),
-        {num_keys_in_block,
+        {(int64_t)num_keys_in_block,
          false ? MatmulGradK::ThreadblockShape::kN : p.head_dim},
         thread_id);
     accumulateInGmem<MatmulGradK>(
@@ -2574,7 +2574,7 @@ struct AttentionBackwardKernel {
       }
     } else {
       int num_iters =
-          ceil_div(p.head_dim_value, kElementsPerAccess * kNumThreadsPerLine) *
+          ceil_div(p.head_dim_value, (int64_t)(kElementsPerAccess * kNumThreadsPerLine)) *
           (kElementsPerAccess * kNumThreadsPerLine);
       for (int iter = 0; iter < num_iters; ++iter) {
         columnIteration(iter);
